// apps/bitrix/modules/konstructor/entities/Complect/type/document-complect-type.ts
export type Complect = {
  name: ComplectNamesEnum; //name for development on english
  fullTitle: ComplectFullTitlesEnum; //Полное официальное наименование - Итоговое , идет в документацию и в название Product и Product Row
  title: ComplectShortitlesEnum; //обычное отображаемое имя
  shortTitle: ComplectShortitlesEnum; //сокращенное наименование для верстки например ГЛ. Бух

  tag: string;
  className: string;
  number: ComplectNumberType;
  weight: ComplectWeightType;

  withConsalting: boolean;
  isChanging: boolean;

  filling: Array<string>;
  ers: Array<number>;
  packetsEr: Array<number>;
  ersInPacket: Array<number>;
  lt: Array<number>;
  ltInPacket: Array<number>;
  withStar: boolean;
  star: Array<number>;
  freeBlocks: Array<number>;
  consalting: Array<number>;
  consaltingProduct: Array<number>;
  type: ComplectsTypes;
  abs?: false | number;
  msk?: number;
  regions?: Array<any>;
};

export type ComplectNumberType =
  | 0
  | 1
  | 2
  | 3
  | 4
  | 5
  | 6
  | 7
  | 8
  | 9
  | 10
  | 11
  | 12
  | 13
  | 14
  | 15
  | 16
  | 17
  | 18;
export type ComplectWeightType =
  | 0
  | 1
  | 1.5
  | 2
  | 2.5
  | 3
  | 3.5
  | 4
  | 6
  | 7
  | 8
  | 9
  | 12
  | 12.5
  | 15.5;
type AbsType = 1 | 2 | 3 | 4 | 4.5 | 5 | 6 | 8 | 10 | 14;

export enum ComplectNamesEnum {
  ProfBuh = "buh",
  ProfBuhgos = "buhgos",
  ProfUr = "ur",
  ProfExpert = "expert",
  ProfOffice = "office",
  ProfGlavbuh = "glavbuh",
  ProfGlavbuhgos = "glavbuhgos",
  ProfExZak = "exzak",

  ProfPred = "company",
  ProfPredPro = "companyPro",
  Classic = "classic",
  ClassicPlus = "classicPlus",
  Universal = "un",
  UniversalPlus = "unPlus",
  Professional = "professional",
  Master = "master",
  Analitik = "analitik",
  AnalitikPlus = "analitikPlus",
  Maximum = "maximum",
  Default = "default",
}
// ...
export enum ComplectFullTitlesEnum {
  ProfBuh = "Гарант-Бухгалтер",
  ProfBuhgos = "Гарант-Бухгалтер госсектора",
  ProfUr = "Гарант-Юрист",
  ProfExpert = "Гарант-Эксперт PRO",
  ProfOffice = "Гарант-Офис",
  ProfExZak = "Эксперт в закупках",

  ProfGlavbuh = "Гарант-Главный Бухгалтер",
  ProfGlavbuhgos = "Гарант-Главный Бухгалтер госсектора",
  ProfPred = "Гарант-Предприятие",
  ProfPredPro = "Гарант-Предприятие PRO",
  Classic = "Гарант-Классик",
  ClassicPlus = "Гарант-Классик+",
  Universal = "Гарант-Универсал",
  UniversalPlus = "Гарант-Универсал+",
  Professional = "Гарант-Профессионал",
  Master = "Гарант-Мастер",
  Analitik = "Гарант-Аналитик",
  AnalitikPlus = "Гарант-Аналитик+",
  Maximum = "Гарант-Максимум",
  Default = "Универсальный комплект",
}

export enum ComplectShortitlesEnum {
  ProfBuh = "Бухгалтер",
  ProfBuhgos = "Бухгалтер госсектора",
  ProfUr = "Юрист",
  ProfExpert = "Эксперт PRO",
  ProfOffice = "Офис",
  ProfExZak = "Эксперт в закупках",

  ProfGlavbuh = "Гл. Бухгалтер",
  ProfGlavbuhgos = "Гл. Бухгалтер госсектора",
  ProfPred = "Предприятие",
  ProfPredPro = "Предприятие PRO",
  Classic = "Классик",
  ClassicPlus = "Классик+",
  Universal = "Универсал",
  UniversalPlus = "Универсал+",
  Professional = "Профессионал",
  Master = "Мастер",
  Analitik = "Аналитик",
  AnalitikPlus = "Аналитик+",
  Maximum = "Максимум",
  Default = "Универсальный комплект",
}
// ...
export enum ComplectsTypes {
  PROF = "prof",
  UNIVERSAL = "universal",
}